+++ saleor/graphql/payment/resolvers.py
@@ -21,4 +21,8 @@ def resolve_payments(info):
 
 
 def resolve_transaction(id):
-    return models.TransactionItem.objects.filter(id=id).first()
+    if id.isdigit():
+        query_params = {"id": id, "use_old_id": True}
+    else:
+        query_params = {"token": id}
+    return models.TransactionItem.objects.filter(**query_params).first()